# openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Mydata API",
		"version": "1.0",
		"description": "API for handling electronic signatures, certification operations, and account information"
	},
	"servers": [
		{
			"url": "http://localhost:5000",
			"description": "Local development server"
		},
		{
			"url": "https://localhost:4200",
			"description": "Other bank API server"
		}
	],
...
	"paths": {
		"/api/oauth/2.0/token": {
			"post": {
				"security": [],
				"summary": "OAuth 2.0 token endpoint",
				"description": "Generates an access token using client credentials grant type",
				"tags": ["OAuth"],
				"operationId": "getIA101",
				"parameters": [
					{
						"in": "header",
						"name": "x-api-tran-id",
						"schema": {
							"type": "string",
							"maxLength": 25,
							"pattern": "^[a-zA-Z0-9]{1,25}$",
							"example": "anya123456SDFRES202503051"
						},
						"required": true,
						"description": "API transaction ID for request tracking"
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/x-www-form-urlencoded": {
							"schema": {
								"type": "object",
								"required": ["grant_type", "client_id", "client_secret", "scope"],
								"properties": {
									"grant_type": {
										"type": "string",
										"enum": ["client_credentials"],
										"description": "OAuth 2.0 grant type (must be 'client_credentials')"
									},
									"client_id": {
										"type": "string",
										"description": "OAuth client ID",
										"maxLength": 50,
										"pattern": "^[a-zA-Z0-9]{1,50}$",
										"example": "wcf8rjy6kphudsnea0l3ytkpdhqrvcxz1578m4q7xv9zb2teee"
									},
									"client_secret": {
										"type": "string",
										"description": "OAuth client secret",
										"maxLength": 50,
										"pattern": "^[a-zA-Z0-9]{1,50}$",
										"example": "wcf8rjy6kphudsnea0l3ytkpdhqrvcxz1578m4q7xv9zb2teee"
									},
									"scope": {
										"type": "string",
										"enum": ["ca"],
										"description": "API access scope (must be 'ca')"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful token generation",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TokenResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request parameters",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Invalid client credentials",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
		"/api/ca/sign_result": {
			"post": {
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Processes signing result and creates signed consent records",
				"description": "Endpoint for handling certificate signing results. Validates input parameters, creates signed consent records from the certificate data, and returns the formatted list.",
				"tags": ["Certification Authority"],
				"operationId": "getIA103",
				"parameters": [
					{
						"in": "header",
						"name": "x-api-tran-id",
						"required": true,
						"schema": {
							"type": "string",
							"maxLength": 25,
							"pattern": "^[a-zA-Z0-9]{1,25}$",
							"example": "anya123456SDFRES202503051"
						},
						"description": "Transaction ID for request tracking"
					},
					{
						"in": "header",
						"name": "Authorization",
						"schema": {
							"type": "string",
							"example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
						},
						"required": true,
						"description": "Bearer token for authorization"
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignResultRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful operation",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SignResultResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad request (Invalid parameters)",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "JWT Authorization header using the Bearer scheme"
			}
		},
		"schemas": {
...
			"SignResultRequest": {
				"type": "object",
				"required": ["cert_tx_id", "sign_tx_id"],
				"properties": {
					"cert_tx_id": {
						"type": "string",
						"description": "Certificate transaction ID",
						"example": "CERT2025001"
					},
					"sign_tx_id": {
						"type": "string",
						"description": "Sign transaction ID",
						"example": "anya123456_certauth00_20250305120000_anyaserial00"
					}
				}
			},
			"SignResultResponse": {
				"type": "object",
				"required": ["rsp_code", "rsp_msg", "signed_consent_list"],
				"properties": {
					"rsp_code": {
						"type": "string",
						"description": "Response code",
						"example": "0000"
					},
					"rsp_msg": {
						"type": "string",
						"description": "Response message",
						"example": "Success"
					},
					"signed_consent_cnt": {
						"type": "integer",
						"description": "Count of signed consents",
						"example": 3
					},
					"signed_consent_list": {
						"type": "array",
						"description": "List of signed consents",
						"items": {
							"$ref": "#/components/schemas/SignedConsent"
						}
					}
				}
			},
			"SignedConsent": {
				"type": "object",
				"required": ["tx_id", "signed_consent", "signed_consent_len"],
				"properties": {
					"tx_id": {
						"type": "string",
						"description": "Transaction ID",
						"example": "MD_anya123456_bond123456_RA20215001_certauth00_20250305120000_XXAB00490001"
					},
					"signed_consent": {
						"type": "string",
						"description": "Signed consent data (base64 encoded)",
						"example": "aHR0cHM6Ly9leGFtcGxlLmNvbQ=="
					},
					"signed_consent_len": {
						"type": "integer",
						"description": "Length of signed consent data",
						"example": 24
					},
					"user_id": {
						"type": "string",
						"description": "User ID",
						"example": "user@example.com"
					},
					"certificate_id": {
						"type": "string",
						"description": "Certificate ID",
						"example": "CERT-123456"
					}
				}
			},
...
			"TokenResponse": {
				"type": "object",
				"required": ["rsp_code", "rsp_msg", "token_type", "access_token", "expires_in", "scope"],
				"properties": {
					"rsp_code": {
						"type": "string",
						"description": "Response code",
						"example": "0000"
					},
					"rsp_msg": {
						"type": "string",
						"description": "Response message",
						"example": "Success"
					},
					"token_type": {
						"type": "string",
						"description": "Token type",
						"example": "Bearer"
					},
					"access_token": {
						"type": "string",
						"description": "Access token",
						"example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
					},
					"expires_in": {
						"type": "integer",
						"description": "Token expiration time in seconds",
						"example": 3600
					},
					"scope": {
						"type": "string",
						"description": "Token scope",
						"example": "ca"
					}
				}
			},
...
			"ErrorResponse": {
				"type": "object",
				"required": ["code", "message"],
				"properties": {
					"code": {
						"type": "string",
						"description": "Error code",
						"example": "ERR_400"
					},
					"message": {
						"type": "string",
						"description": "Error message",
						"example": "Invalid request parameters"
					}
				}
			}
		}
	}
}